+++ strategy/from_json/to_serializable.ts
@@ -2,8 +2,7 @@
 
 import { JSONValue, Serializable } from "../../serializable.ts";
 import { FromJSONStrategy } from "../compose_strategy.ts";
-import { StrategyTypeArgument } from "../_utils.ts";
-import { getNew } from "../utils.ts";
+import { getNew, StrategyTypeArgument } from "../utils.ts";
 
 /** revive data using `fromJSON` on a subclass type */
 export function toSerializable<T>(