+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- docs for @SerializeProperty short cut functions   
 - new `strategy/utils.ts` exposing `getNew(type)` for custom decorators
 
 ## [v1.1.4] - 2021-01-19
@@ -21,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - deno@1.6.3 fixing #96
 - std@0.83.0
 
-## [v1.1.2] -2021-01-08
+## [v1.1.2] - 2021-01-08
 
 ### Changed
 - fixed a typo in nodeJS export